backend/pyproject.toml
[project]
name = "eloquent-ai-chatbot-backend"
version = "1.0.0"
description = "FastAPI backend for Eloquent AI Chatbot with RAG"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "python-multipart==0.0.6",
    "slowapi==0.1.9",
    "motor==3.3.2",
    "pymongo==4.6.1",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.1.2",
    "openai==1.12.0",
    "pinecone-client==3.0.2",
    "tiktoken==0.6.0",
    "python-dotenv==1.0.1",
    "pydantic==2.6.0",
    "pydantic-settings==2.1.0",
    "email-validator==2.2.0",
    "python-dateutil==2.8.2",
    "httpx==0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
